+++ color/src/serialize.rs
@@ -86,6 +86,7 @@ impl core::fmt::Display for DynamicColor {
             ColorSpaceTag::A98Rgb => write_color_function(self, "a98-rgb", f),
             ColorSpaceTag::ProphotoRgb => write_color_function(self, "prophoto-rgb", f),
             ColorSpaceTag::Rec2020 => write_color_function(self, "rec2020", f),
+            ColorSpaceTag::AcesCg => write_color_function(self, "--acescg", f),
             ColorSpaceTag::Hsl => write_legacy_function(self, "hsl", 1.0, f),
             ColorSpaceTag::Hwb => write_modern_function(self, "hwb", f),
             ColorSpaceTag::XyzD50 => write_color_function(self, "xyz-d50", f),