+++ color/src/tag.rs
@@ -4,8 +4,8 @@
 //! The color space tag enum.
 
 use crate::{
-    A98Rgb, ColorSpace, ColorSpaceLayout, DisplayP3, Hsl, Hwb, Lab, Lch, LinearSrgb, Missing,
-    Oklab, Oklch, ProphotoRgb, Rec2020, Srgb, XyzD50, XyzD65,
+    A98Rgb, AcesCg, ColorSpace, ColorSpaceLayout, DisplayP3, Hsl, Hwb, Lab, Lch, LinearSrgb,
+    Missing, Oklab, Oklch, ProphotoRgb, Rec2020, Srgb, XyzD50, XyzD65,
 };
 
 /// The color space tag for dynamic colors.
@@ -44,6 +44,8 @@ pub enum ColorSpaceTag {
     ProphotoRgb,
     /// The [`Rec2020`] color space.
     Rec2020,
+    /// The [`AcesCg`] color space.
+    AcesCg,
     /// The [`XyzD50`] color space.
     XyzD50,
     /// The [`XyzD65`] color space.
@@ -67,8 +69,22 @@ impl ColorSpaceTag {
         matches!(
             (self, other),
             (
-                Srgb | LinearSrgb | DisplayP3 | A98Rgb | ProphotoRgb | Rec2020 | XyzD50 | XyzD65,
-                Srgb | LinearSrgb | DisplayP3 | A98Rgb | ProphotoRgb | Rec2020 | XyzD50 | XyzD65
+                Srgb | LinearSrgb
+                    | DisplayP3
+                    | A98Rgb
+                    | ProphotoRgb
+                    | Rec2020
+                    | AcesCg
+                    | XyzD50
+                    | XyzD65,
+                Srgb | LinearSrgb
+                    | DisplayP3
+                    | A98Rgb
+                    | ProphotoRgb
+                    | Rec2020
+                    | AcesCg
+                    | XyzD50
+                    | XyzD65
             ) | (Lab | Oklab, Lab | Oklab)
                 | (Lch | Oklch, Lch | Oklch)
         )
@@ -145,6 +161,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::from_linear_srgb(rgb),
             Self::ProphotoRgb => ProphotoRgb::from_linear_srgb(rgb),
             Self::Rec2020 => Rec2020::from_linear_srgb(rgb),
+            Self::AcesCg => AcesCg::from_linear_srgb(rgb),
             Self::XyzD50 => XyzD50::from_linear_srgb(rgb),
             Self::XyzD65 => XyzD65::from_linear_srgb(rgb),
             Self::Hsl => Hsl::from_linear_srgb(rgb),
@@ -167,6 +184,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::to_linear_srgb(src),
             Self::ProphotoRgb => ProphotoRgb::to_linear_srgb(src),
             Self::Rec2020 => Rec2020::to_linear_srgb(src),
+            Self::AcesCg => AcesCg::to_linear_srgb(src),
             Self::XyzD50 => XyzD50::to_linear_srgb(src),
             Self::XyzD65 => XyzD65::to_linear_srgb(src),
             Self::Hsl => Hsl::to_linear_srgb(src),
@@ -223,6 +241,7 @@ impl ColorSpaceTag {
             Self::A98Rgb => A98Rgb::clip(src),
             Self::ProphotoRgb => ProphotoRgb::clip(src),
             Self::Rec2020 => Rec2020::clip(src),
+            Self::AcesCg => AcesCg::clip(src),
             Self::XyzD50 => XyzD50::clip(src),
             Self::XyzD65 => XyzD65::clip(src),
             Self::Hsl => Hsl::clip(src),